Move `Public` prop to the base `ProjectRef`
in order to consume the project.public from `Repository` holds a `ProjectRef` and not `Project` (though it seems to actually return something more similar to `Project` from the raw rest api).

I could change the Repository to hold Project instead, but it's being used all over, I wanted to only add this specific field and not the entire model.

but I think for the long run, we should ditch this package completely and just implement our own rest-api client based on the bitbucketserver docs.